Removed other handle new value and node class, updated every component to use the same one
