--- src/frontend/src/modals/editNodeModal/hooks/use-handle-node-class.tsx
@@ -1,32 +1,0 @@
-import { cloneDeep } from "lodash";
-
-const useHandleNodeClass = (
-  takeSnapshot: () => void,
-  setNode: (id: string, callback: (oldNode: any) => any) => void,
-  nodeId: string,
-) => {
-  const handleNodeClass = (newNodeClass, name, code, type?: string) => {
-    if (code) {
-      takeSnapshot();
-    }
-
-    setNode(nodeId, (oldNode) => {
-      let newNode = cloneDeep(oldNode);
-
-      newNode.data = {
-        ...newNode.data,
-        node: cloneDeep(newNodeClass),
-      };
-      if (type) {
-        newNode.data.node.template[name].type = type;
-      }
-      newNode.data.node.template[name].value = code;
-
-      return newNode;
-    });
-  };
-
-  return { handleNodeClass };
-};
-
-export default useHandleNodeClass;
