+++ src/frontend/src/modals/editNodeModal/index.tsx
@@ -1,6 +1,7 @@
 import { ColDef } from "ag-grid-community";
 import { forwardRef, useState } from "react";
 import { useUpdateNodeInternals } from "reactflow";
+import useHandleNodeClass from "../../CustomNodes/hooks/use-handle-node-class";
 import TableComponent from "../../components/tableComponent";
 import { Badge } from "../../components/ui/badge";
 import { Button } from "../../components/ui/button";
@@ -12,7 +13,6 @@ import { NodeDataType } from "../../types/flow";
 import BaseModal from "../baseModal";
 import useColumnDefs from "./hooks/use-column-defs";
 import useHandleChangeAdvanced from "./hooks/use-handle-change-advanced";
-import useHandleNodeClass from "./hooks/use-handle-node-class";
 import useRowData from "./hooks/use-row-data";
 
 const EditNodeModal = forwardRef(
